store: add List to store interface
Stealth needs a `List` method in order to be able to list all secrets
for a Namespace (Env + App). This is needed to support the CLI tool in
https://clever.atlassian.net/browse/INFRA-1886

---

note: the Unicreds implementation currently relies on ListAll. This
currently queries all keys... there is no filter on the dyanmodb query.

Since this is also how we do `History`, I'll stick with this for now and
we can make the upstream changes to unicreds later.