bug: Ensure that update is only called if secret already exists
When using stealth directly with Parameter Store, the function createOrUpdate()
will try to first create the secret, and if an error is returned, validate that
the error is NOT that the secret exists. Then, it'll go through with updating
the secret.

There's a case where the secret has not yet been created. This causes Create()
to return no error, which then causes Update() to be called as well.

By ensuring that Create() has returned that the secret exists, Update() should
then be called after.

This should solve `./stealth write` from writing two versions of a newly created
secret.